+++ src/utils/is-prod.ts
@@ -1,0 +1,4 @@
+const { NODE_ENV } = process.env;
+const isProd = NODE_ENV === 'production';
+
+export default isProd;
